mammys_freezer_api/requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
et-xmlfile==1.1.0
Flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
numpy==1.19.5
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
xlrd==2.0.1
zipp==3.4.1
